Change qemu path to system default
This tool was designed for SPP and it was convenience to fixed path
for customized qemu. However, it is better to be the defalt is the
path is system default for general purpose.
The path of qemu is configured with -q or --qemu option if you use
other than default.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>