Change directory structure
This update is for clear directory structure of hda files.
run-vm.py is moved from runscripts/ to /bin and runscripts/ is renamed
to hda/.
Original hda is also moved from iso/ to hda/ for maintainance easily.

Add hda/README.md for describing the directory structure.

TODO: run-vm.py is still working for old dir structure and it is needed
to update for new one.

Signed-off-by: Yasufumi Ogawa <ogawa.yasufumi@lab.ntt.co.jp>